+++ src/DI/firewallServices.neon
@@ -1,4 +1,4 @@
 services:
 	- Symfony\Component\Security\Http\FirewallMap
-	- Symnedi\Security\Http\Firewall
+	- Symnedi\Security\EventSubscriber\FirewallSubscriber
 	- Symnedi\Security\Bridge\SymfonyHttpFoundation\Request\SymfonyRequestAdapterFactory
